move setOrientation implementation to simulator
